requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asgiref==3.7.2
python-dotenv==1.0.0
